[CG] Use uint64_t for the return type of getArrayNumElements.
Fix a potential issue for gather2loadpermute optimization,
when array element number is bigger than 0x100000000,
and the last 32 bits is less than 8, it may be wrong.

Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/296623
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: Pei, Gen <gen.pei@intel.com>

+++ llvm/include/llvm/Analysis/TargetTransformInfoImpl.h
@@ -240,7 +240,7 @@ public:
     return !isLegalMaskedGather(LoadTy, Alignment);
   }
 
-  bool shouldOptGatherToLoadPermute(Type *ArrayElemTy, uint32_t ArrayNum,
+  bool shouldOptGatherToLoadPermute(Type *ArrayElemTy, uint64_t ArrayNum,
                                     uint32_t GatherNum,
                                     uint32_t *WidenNum = nullptr) const {
     return false;
@@ -248,7 +248,7 @@ public:
 
   bool isLegalToTransformGather2PermuteLoad(const IntrinsicInst *II,
                                             Type *&ArrayElemTy,
-                                            unsigned &ArrayNum,
+                                            uint64_t &ArrayNum,
                                             unsigned &GatherNum,
                                             unsigned &WidenNum) const {
     return false;
@@ -256,7 +256,7 @@ public:
 
   bool isLegalToTransformGather2PermuteLoad(
       Intrinsic::ID ID, Type *DataTy, const Value *Ptr, bool VariableMask,
-      bool UndefPassThru, Type *&ArrayElemTy, unsigned &ArrayNum,
+      bool UndefPassThru, Type *&ArrayElemTy, uint64_t &ArrayNum,
       unsigned &GatherNum, unsigned &WidenNum) const {
     return false;
   }
